+++ uui-editor/src/migrations/slate_migrations.ts
@@ -1,12 +1,11 @@
 import { TDescendant, Value } from '@udecode/plate-common';
-import { toPlateAlign } from '../helpers';
 import { IFRAME_TYPE } from '../plugins/iframePlugin/constants';
 import { IMAGE_TYPE } from '../plugins/imagePlugin/constants';
+import { TImageElement } from '../plugins/imagePlugin/types';
 import { LINK_TYPE } from '../plugins/linkPlugin/constants';
 import { PARAGRAPH_TYPE } from '../plugins/paragraphPlugin';
 import { TABLE_CELL_TYPE, TABLE_HEADER_CELL_TYPE, TABLE_ROW_TYPE, TABLE_TYPE } from '../plugins/tablePlugin/constants';
-import { EditorValue } from '../types';
-import { SlateBlockElement, SlateElement, SlateInlineElement, SlateMark, SlateSchema, SlateTextElement } from './types';
+import { SlateBlockElement, SlateElement, SlateImgAlign, SlateInlineElement, SlateMark, SlateSchema, SlateTextElement } from './types';
 
 const mediaTypes = [IMAGE_TYPE, IFRAME_TYPE];
 const cellTypes = [TABLE_CELL_TYPE, TABLE_HEADER_CELL_TYPE];
@@ -32,8 +31,10 @@ const migrateTextNode = (oldNode: SlateTextElement) => {
 const getTablePayload = (node: SlateElement) => {
     if (node.type !== TABLE_TYPE) return {};
 
-    if (node.data.cellSizes) {
-        return { colSizes: [...node.data.cellSizes] };
+    const { cellSizes, ...otherData } = node.data || {};
+    if (!!cellSizes) {
+        const pyaload = { colSizes: [...cellSizes], data: { ...otherData } };
+        return pyaload;
     }
 
     return {};
@@ -59,26 +60,29 @@ const getTableRowPayload = (node: SlateElement): { children: TDescendant[] } => 
 const getTableCellElementPayload = (node: SlateElement) => {
     if (!cellTypes.includes(node.type || '')) return {};
 
-    const colSpanPayload = node.data?.colSpan ? { colSpan: node.data.colSpan } : {};
-    const rowSpanPayload = node.data?.rowSpan ? { rowSpan: node.data.rowSpan } : {};
+    const { colSpan, rowSpan, ...otherData } = node.data || {};
+    const colSpanPayload = colSpan ? { colSpan } : {};
+    const rowSpanPayload = rowSpan ? { rowSpan } : {};
 
-    return { ...colSpanPayload, ...rowSpanPayload };
+    return { ...colSpanPayload, ...rowSpanPayload, data: { ...otherData } };
 };
 
 const getLinkPayload = (node: SlateElement) => {
     if (node.type !== LINK_TYPE) return {};
 
-    return node.data?.url ? { url: node.data.url } : {};
+    const { url, ...otherData } = node.data || {};
+    return url ? { url, data: { ...otherData } } : {};
 };
 
 /** iframes and images */
 const getMediaTypesPayload = (node: SlateElement) => {
     if (!mediaTypes.includes(node.type || '')) return {};
 
-    const urlPayload = node.data?.src ? { url: node.data.src } : {};
-    const alignPayload = node.data?.align ? { align: toPlateAlign(node.data.align) } : {};
+    const { src, align, ...other } = node.data || {};
+    const urlPayload = src ? { url: src } : {};
+    const alignPayload = align ? { align: toPlateAlign(align) } : {};
 
-    return { ...urlPayload, ...alignPayload };
+    return { ...urlPayload, ...alignPayload, data: { ...other } };
 };
 
 const isTable = (node: SlateElement): boolean => {
@@ -101,11 +105,11 @@ const migrateDeeper = (_node: SlateBlockElement | SlateInlineElement): TDescenda
     /** there might be duplications with data, but we could also omit them with versioned apprach via plate normalizeNode */
     return {
         type: node.type,
+        ...dataPayload,
         ...getTablePayload(node),
         ...getTableCellElementPayload(node),
         ...getLinkPayload(node),
         ...getMediaTypesPayload(node),
-        ...dataPayload,
         ...childrenPayload,
     };
 };
@@ -131,7 +135,12 @@ export const migrateSlateSchema = (schema: SlateSchema): Value => {
     return schema as unknown as Value;
 };
 
-/** type guard to distinct slate format */
-export const isSlateSchema = (value: EditorValue): value is SlateSchema => {
-    return !!value && !Array.isArray(value);
+const SLATE_TO_PLATE_IMG_ALIGN = {
+    'align-left': 'left',
+    'align-right': 'right',
+    'align-center': 'center',
 };
+
+/** converts align property */
+export const toPlateAlign = (slateAlign: SlateImgAlign): TImageElement['align'] =>
+    SLATE_TO_PLATE_IMG_ALIGN[slateAlign] as TImageElement['align'];