+++ uui-editor/src/plugins/imagePlugin/ImageBlock.tsx
@@ -17,7 +17,7 @@ import { ImageElement } from './ImageElement';
 
 import css from './ImageBlock.module.scss';
 import { ImgToolbar } from './Toolbar';
-import { IImageElement, PlateImgAlign } from './types';
+import { PlateImgAlign, TImageElement } from './types';
 import { useMediaState } from '@udecode/plate-media';
 import { IMAGE_TYPE } from './constants';
 
@@ -26,7 +26,7 @@ const IMAGE_STYLES = {
     paddingBottom: 0,
 };
 
-export const Image: PlatePluginComponent<PlateRenderElementProps<Value, IImageElement>> = function (props) {
+export const Image: PlatePluginComponent<PlateRenderElementProps<Value, TImageElement>> = function (props) {
     const {
         editor, element, children,
     } = props;